creates placeApiImpl.go, redisImpl.go and stateMachine.go
