+++ datahub/controllers/datahub.go
@@ -134,189 +134,3 @@ func (h DataHubController) handleSession() { //sub to channel, continuously re p
 	}
 
 }
-
-func (h *DataHubController) handleCases(c ClientPayload) (DataHubPayload, bool, bool) {
-
-	datahubPayload := DataHubPayload{}
-	closeConnection := false
-	errorVal := false
-
-	switch c.RequestType {
-	case "leaveSession":
-		delete(h.currentUserIDs, c.ClientID)
-		if len(h.currentUserIDs) == 0 {
-			closeConnection = true
-			h.cleanRedisDB()
-		}
-		log.Println(h.currentUserIDs)
-		datahubPayload.State = UserStatus.String(3)
-
-	case "joinSession":
-		h.currentUserIDs[c.ClientID] = Idle
-		log.Println(h.currentUserIDs)
-		datahubPayload.State = UserStatus.String(0)
-
-	case "updateRestaurants":
-		datahubPayload.State = UserStatus.String(4)
-		if h.currentUserIDs[c.ClientID] != UpdateRestaurants {
-			h.updateRestaurantsCounter += 1
-			h.currentUserIDs[c.ClientID] = UpdateRestaurants
-		}
-
-		if h.updateRestaurantsCounter == len(h.currentUserIDs) {
-			for key := range h.currentUserIDs {
-				h.currentUserIDs[key] = CurrRating
-			}
-
-			h.updateRestaurantsCounter = 0
-			datahubPayload.State = UserStatus.String(4)
-			datahubPayload.PlaceApiData = h.getNewRestaurants()
-
-			h.initializeRedisDB()
-		}
-
-	case "startRating":
-		datahubPayload.State = UserStatus.String(1)
-		if h.currentUserIDs[c.ClientID] != StartRating {
-			h.startRatingCounter += 1
-			h.currentUserIDs[c.ClientID] = StartRating
-		}
-
-		if h.startRatingCounter == len(h.currentUserIDs) {
-			for key := range h.currentUserIDs {
-				h.currentUserIDs[key] = CurrRating
-			}
-
-			h.startRatingCounter = 0
-			datahubPayload.State = UserStatus.String(2)
-			datahubPayload.PlaceApiData = h.getNewRestaurants()
-
-			h.initializeRedisDB()
-		}
-
-	case "finishRating":
-		datahubPayload.State = UserStatus.String(3)
-		if h.currentUserIDs[c.ClientID] != FinishRating {
-			h.finishRatingCounter += 1
-			h.currentUserIDs[c.ClientID] = FinishRating
-		}
-
-		if h.finishRatingCounter == len(h.currentUserIDs) {
-			datahubPayload.ResultsData.SearchResult = append(datahubPayload.ResultsData.SearchResult, h.getRatingResult())
-			log.Println("FINISHED RATING -----------------------")
-		}
-
-	case "sendResult":
-		datahubPayload.State = UserStatus.String(2)
-		h.updateScore(c.RestaurantID)
-
-	default:
-		closeConnection = true
-		errorVal = true
-	}
-
-	return datahubPayload, closeConnection, errorVal
-}
-
-func (h *DataHubController) getPlaceAPIData() maps.PlacesSearchResponse {
-	r := &maps.NearbySearchRequest{
-		Location: &maps.LatLng{
-			Lat: 43.475074,
-			Lng: -80.543213,
-		},
-		Radius:    10000,
-		OpenNow:   true,
-		Type:      maps.PlaceTypeRestaurant,
-		RankBy:    maps.RankByProminence,
-		PageToken: h.placeApiData.NextPageToken,
-	}
-
-	response, err := h.mapsClient.NearbySearch(context.Background(), r)
-	if err != nil {
-		log.Fatalf("fatal error: %s", err)
-		return maps.PlacesSearchResponse{}
-	}
-
-	return response
-}
-
-func (h *DataHubController) getNewRestaurants() maps.PlacesSearchResponse {
-	if len(h.placeApiData.Results) == 0 {
-		h.placeApiData = h.getPlaceAPIData()
-		log.Println("had to do new api call, len results: ", len(h.placeApiData.Results))
-	}
-
-	searchResponse := h.placeApiData
-	searchResponse.Results = searchResponse.Results[:9] // get only first ten results
-
-	h.placeApiData.Results = h.placeApiData.Results[10:] // remove first 10 results
-
-	return searchResponse
-}
-
-func (h *DataHubController) initializeRedisDB() {
-	ctx := context.Background()
-
-	h.cleanRedisDB()
-
-	for _, restaurant := range h.placeApiData.Results {
-		err := h.redisClient.ZAdd(ctx, h.sessionID.String()+"set", &redis.Z{
-			Score:  0,
-			Member: restaurant.PlaceID,
-		}).Err()
-		if err != nil {
-			panic(err)
-		}
-
-		marshaledRestaurant, err := json.Marshal(restaurant)
-		if err != nil {
-			panic(err)
-		}
-
-		err = h.redisClient.HSet(ctx, h.sessionID.String()+"hash", restaurant.PlaceID, string(marshaledRestaurant)).Err()
-		if err != nil {
-			panic(err)
-		}
-	}
-}
-
-func (h *DataHubController) cleanRedisDB() {
-	ctx := context.Background()
-
-	err := h.redisClient.Del(ctx, h.sessionID.String()+"set", h.sessionID.String()+"hash").Err()
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (h *DataHubController) updateScore(PlaceID string) {
-	ctx := context.Background()
-
-	err := h.redisClient.ZIncrBy(ctx, h.sessionID.String()+"set", 1, PlaceID).Err()
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (h *DataHubController) getRatingResult() maps.PlacesSearchResult {
-	ctx := context.Background()
-
-	key, err := h.redisClient.ZRevRange(ctx, h.sessionID.String()+"set", 0, 0).Result()
-	if err != nil {
-		panic(err)
-	}
-
-	result, err := h.redisClient.HGet(ctx, h.sessionID.String()+"hash", key[0]).Result()
-	if err != nil {
-		log.Println(err)
-	}
-
-	marshaledResult := maps.PlacesSearchResult{}
-	err = json.Unmarshal([]byte(result), &marshaledResult)
-	if err != nil {
-		log.Println(err)
-	}
-
-	return marshaledResult
-
-}
